{"types":"C","runs":12}

+++ BuildHouse.java
@@ -19,7 +19,7 @@ public class BuildHouse extends GraphicsProgram {
 
 	private void drawWall() {
 		GRect wall = new GRect (HOUSE_HEIGHT, HOUSE_WIDTH);
-		int x =(int) (getWidth() / 2 - HOUSE_WIDTH / 2);
+		double x = (getWidth() / 2 - HOUSE_WIDTH / 2);
 		double y = getHeight() - HOUSE_HEIGHT;
 		
 		add(wall, x, y);
